First whack at DB Schema
This is a first whack at getting a fully backwards-compatible database schema. It was assembled by looking at the previous few years (mostly 2009)--any field that was not entirely NULL was ported over. There are a few fields that I'm unsure of (see the TODO file), but those can come later