First whack at registration forms
I mocked out a 3-part registration process. First you register an
account, then your school, then the teams that are competing with your
school.

The previous registration combined steps 1 and 2 in this flow. I want to
seperate them so we can have a many-to-{one|many} relationship for
schools and coaches (= accounts). This seems like the Right Thing To Do.

Steps 2 and 3 are distinct because otherwise there would be tons of
unrelated information on one page, which is no good.

And I'm aware the UI sucks right now.